// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OperationType = {
  venta: 'venta',
  compra: 'compra'
} as const

export type OperationType = (typeof OperationType)[keyof typeof OperationType]


export const Role = {
  admin: 'admin',
  vendedor: 'vendedor'
} as const

export type Role = (typeof Role)[keyof typeof Role]